Trying out seperating routes and models on the backend
